feat: extract `createImageUrl` into util file
